Move `*ByteSlice*` traits into `byte_slice` module
This helps declutter our top-level documentation.
